requirements.txt
google-adk>=1.0.0
pydantic>=2.0.0
